requirements.txt
# Requires Python 3.10 or higher
mcp>=1.6.0
google-auth>=2.39.0
google-auth-oauthlib>=1.2.1
google-api-python-client>=2.167.0
python-dateutil>=2.9.0
dateparser
openai==2.0.0
openai-agents
llama-index
llama-index-tools-mcp
# phoenix:
openinference-instrumentation-openai
arize-phoenix-otel>=0.4.0
opentelemetry-sdk==1.27.0
opentelemetry-exporter-otlp-proto-http==1.27.0
opentelemetry-instrumentation==0.48b0
neo4j # not required, but added for consistency